+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Fixes:
 - Some expressions that included many extractable sub-expressions (map literals, blocking calls) generated invalid output because a temporary variable was re-used incorrectly
 - Empty statements are accepted at the top level
 - Ignore extra arguments in blocking function calls in all cases. Previously, extra arguments were ignored in some cases but threw errors in other cases.
+- Call expressions in optional chaining (`data?.getPerson()?.name`) now result in error. Previously, they were incorrectly treated as non-optional in the generated code.
 
 ## Version 0.11.0 - 2026-06-28