+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # ts2workflows changelog
 
-## unreleased
+## Version 0.13.0 - 2025-10-12
 
 Fixes: